ProductsRoutes and ProductsController

+++ src/database/knex/migrations/20230822053234_createProduct.js
@@ -3,7 +3,7 @@ exports.up = knex => knex.schema.createTable("products", table => {
     table.text("name");
     table.text("type");
     table.text("description");
-    table.integer("price");
+    table.float("price");
     table.text("avatar").default(null);
 
     table.timestamp("created_at").default(knex.fn.now());
